Move tests into separate directory
This moves most of the tests into the top level integration tests directory. This makes it more obvious that we actually have tests!

It should also make it easier to see what tests rely on private features that have not been exposed. I didn’t move the lexer tests for example, because they rely on the private Lexer type.
