+++ src/lib.rs
@@ -129,9 +129,6 @@ extern crate codespan;
 extern crate codespan_reporting;
 #[macro_use]
 extern crate failure;
-#[cfg(test)]
-extern crate goldenfile;
-#[cfg_attr(test, macro_use)]
 extern crate im;
 extern crate lalrpop_util;
 #[macro_use]
