+++ tests/check.rs
@@ -1,4 +1,14 @@
-use super::*;
+extern crate codespan;
+extern crate codespan_reporting;
+extern crate moniker;
+extern crate pikelet;
+
+use codespan::CodeMap;
+
+use pikelet::semantics::{self, TcEnv, TypeError};
+use pikelet::syntax::translation::{Desugar, DesugarEnv};
+
+mod support;
 
 #[test]
 fn record() {
@@ -8,8 +18,8 @@ fn record() {
     let expected_ty = r"Record { t : Type; x : String }";
     let given_expr = r#"record { t = String; x = "hello" }"#;
 
-    let expected_ty = parse_nf_term(&mut codemap, &tc_env, expected_ty);
-    parse_check_term(&mut codemap, &tc_env, given_expr, &expected_ty);
+    let expected_ty = support::parse_nf_term(&mut codemap, &tc_env, expected_ty);
+    support::parse_check_term(&mut codemap, &tc_env, given_expr, &expected_ty);
 }
 
 #[test]
@@ -21,12 +31,12 @@ fn record_field_mismatch_lt() {
     let expected_ty = r"Record { x : String; y : String }";
     let given_expr = r#"record { x = "hello" }"#;
 
-    let expected_ty = parse_nf_term(&mut codemap, &tc_env, expected_ty);
-    let raw_term = parse_term(&mut codemap, given_expr)
+    let expected_ty = support::parse_nf_term(&mut codemap, &tc_env, expected_ty);
+    let raw_term = support::parse_term(&mut codemap, given_expr)
         .desugar(&desugar_env)
         .unwrap();
 
-    match check_term(&tc_env, &raw_term, &expected_ty) {
+    match semantics::check_term(&tc_env, &raw_term, &expected_ty) {
         Err(TypeError::RecordSizeMismatch { .. }) => {},
         Err(err) => panic!("unexpected error: {:?}", err),
         Ok(term) => panic!("expected error but found: {}", term),
@@ -42,12 +52,12 @@ fn record_field_mismatch_gt() {
     let expected_ty = r"Record { x : String }";
     let given_expr = r#"record { x = "hello"; y = "hello" }"#;
 
-    let expected_ty = parse_nf_term(&mut codemap, &tc_env, expected_ty);
-    let raw_term = parse_term(&mut codemap, given_expr)
+    let expected_ty = support::parse_nf_term(&mut codemap, &tc_env, expected_ty);
+    let raw_term = support::parse_term(&mut codemap, given_expr)
         .desugar(&desugar_env)
         .unwrap();
 
-    match check_term(&tc_env, &raw_term, &expected_ty) {
+    match semantics::check_term(&tc_env, &raw_term, &expected_ty) {
         Err(TypeError::RecordSizeMismatch { .. }) => {},
         Err(err) => panic!("unexpected error: {:?}", err),
         Ok(term) => panic!("expected error but found: {}", term),
@@ -62,8 +72,8 @@ fn dependent_record() {
     let expected_ty = r"Record { t : Type; x : t }";
     let given_expr = r#"record { t = String; x = "hello" }"#;
 
-    let expected_ty = parse_nf_term(&mut codemap, &tc_env, expected_ty);
-    parse_check_term(&mut codemap, &tc_env, given_expr, &expected_ty);
+    let expected_ty = support::parse_nf_term(&mut codemap, &tc_env, expected_ty);
+    support::parse_check_term(&mut codemap, &tc_env, given_expr, &expected_ty);
 }
 
 #[test]
@@ -74,8 +84,8 @@ fn dependent_record_propagate_types() {
     let expected_ty = r"Record { t : Type; x : t }";
     let given_expr = r#"record { t = S32; x = 1 }"#;
 
-    let expected_ty = parse_nf_term(&mut codemap, &tc_env, expected_ty);
-    parse_check_term(&mut codemap, &tc_env, given_expr, &expected_ty);
+    let expected_ty = support::parse_nf_term(&mut codemap, &tc_env, expected_ty);
+    support::parse_check_term(&mut codemap, &tc_env, given_expr, &expected_ty);
 }
 
 #[test]
@@ -90,8 +100,8 @@ fn case_expr() {
         greeting => (import "prim/string/append") greeting "!!";
     }"#;
 
-    let expected_ty = parse_nf_term(&mut codemap, &tc_env, expected_ty);
-    parse_check_term(&mut codemap, &tc_env, given_expr, &expected_ty);
+    let expected_ty = support::parse_nf_term(&mut codemap, &tc_env, expected_ty);
+    support::parse_check_term(&mut codemap, &tc_env, given_expr, &expected_ty);
 }
 
 #[test]
@@ -106,12 +116,12 @@ fn case_expr_bad_literal() {
         1 => "byee";
     }"#;
 
-    let expected_ty = parse_nf_term(&mut codemap, &tc_env, expected_ty);
-    let raw_term = parse_term(&mut codemap, given_expr)
+    let expected_ty = support::parse_nf_term(&mut codemap, &tc_env, expected_ty);
+    let raw_term = support::parse_term(&mut codemap, given_expr)
         .desugar(&desugar_env)
         .unwrap();
 
-    match check_term(&tc_env, &raw_term, &expected_ty) {
+    match semantics::check_term(&tc_env, &raw_term, &expected_ty) {
         Err(TypeError::LiteralMismatch { .. }) => {},
         Err(err) => panic!("unexpected error: {:?}", err),
         Ok(term) => panic!("expected error but found: {}", term),
@@ -128,8 +138,8 @@ fn case_expr_wildcard() {
         _ => 123;
     }"#;
 
-    let expected_ty = parse_nf_term(&mut codemap, &tc_env, expected_ty);
-    parse_check_term(&mut codemap, &tc_env, given_expr, &expected_ty);
+    let expected_ty = support::parse_nf_term(&mut codemap, &tc_env, expected_ty);
+    support::parse_check_term(&mut codemap, &tc_env, given_expr, &expected_ty);
 }
 
 #[test]
@@ -140,8 +150,8 @@ fn case_expr_empty() {
     let expected_ty = r"String";
     let given_expr = r#"case "helloo" {}"#;
 
-    let expected_ty = parse_nf_term(&mut codemap, &tc_env, expected_ty);
-    parse_check_term(&mut codemap, &tc_env, given_expr, &expected_ty);
+    let expected_ty = support::parse_nf_term(&mut codemap, &tc_env, expected_ty);
+    support::parse_check_term(&mut codemap, &tc_env, given_expr, &expected_ty);
 }
 
 #[test]
@@ -152,8 +162,8 @@ fn array_0_string() {
     let expected_ty = r"Array 0 String";
     let given_expr = r#"[]"#;
 
-    let expected_ty = parse_nf_term(&mut codemap, &tc_env, expected_ty);
-    parse_check_term(&mut codemap, &tc_env, given_expr, &expected_ty);
+    let expected_ty = support::parse_nf_term(&mut codemap, &tc_env, expected_ty);
+    support::parse_check_term(&mut codemap, &tc_env, given_expr, &expected_ty);
 }
 
 #[test]
@@ -164,8 +174,8 @@ fn array_3_string() {
     let expected_ty = r"Array 3 String";
     let given_expr = r#"["hello"; "hi"; "byee"]"#;
 
-    let expected_ty = parse_nf_term(&mut codemap, &tc_env, expected_ty);
-    parse_check_term(&mut codemap, &tc_env, given_expr, &expected_ty);
+    let expected_ty = support::parse_nf_term(&mut codemap, &tc_env, expected_ty);
+    support::parse_check_term(&mut codemap, &tc_env, given_expr, &expected_ty);
 }
 
 #[test]
@@ -177,12 +187,12 @@ fn array_len_mismatch() {
     let expected_ty = r"Array 3 String";
     let given_expr = r#"["hello"; "hi"]"#;
 
-    let expected_ty = parse_nf_term(&mut codemap, &tc_env, expected_ty);
-    let raw_term = parse_term(&mut codemap, given_expr)
+    let expected_ty = support::parse_nf_term(&mut codemap, &tc_env, expected_ty);
+    let raw_term = support::parse_term(&mut codemap, given_expr)
         .desugar(&desugar_env)
         .unwrap();
 
-    match check_term(&tc_env, &raw_term, &expected_ty) {
+    match semantics::check_term(&tc_env, &raw_term, &expected_ty) {
         Err(TypeError::ArrayLengthMismatch { .. }) => {},
         Err(err) => panic!("unexpected error: {:?}", err),
         Ok(term) => panic!("expected error but found: {}", term),
@@ -198,12 +208,12 @@ fn array_elem_ty_mismatch() {
     let expected_ty = r"Array 3 String";
     let given_expr = r#"["hello"; "hi"; 4]"#;
 
-    let expected_ty = parse_nf_term(&mut codemap, &tc_env, expected_ty);
-    let raw_term = parse_term(&mut codemap, given_expr)
+    let expected_ty = support::parse_nf_term(&mut codemap, &tc_env, expected_ty);
+    let raw_term = support::parse_term(&mut codemap, given_expr)
         .desugar(&desugar_env)
         .unwrap();
 
-    match check_term(&tc_env, &raw_term, &expected_ty) {
+    match semantics::check_term(&tc_env, &raw_term, &expected_ty) {
         Err(_) => {},
         Ok(term) => panic!("expected error but found: {}", term),
     }
