+++ tests/goldenfiles/ann_ann_left
