+++ tests/pikelet.rs
@@ -1,0 +1,8 @@
+extern crate pikelet;
+
+use pikelet::Pikelet;
+
+#[test]
+fn with_prelude() {
+    let _pikelet = Pikelet::with_prelude();
+}
